refactor: Renamed 'utilities/utils' to 'utilities/helloworld'.
