+++ jutl/utilities/__init__.py
@@ -1,8 +1,8 @@
 # Import utils.py so module
 # functions are usable at
 # 'from jutl import utilities' level.
-from .utils import *
+from .helloworld import *
 
 # Only show functions specified in
 # submodule files to the outside world.
-__all__ = utils.__all__
+__all__ = helloworld.__all__
